refactor: move task response abi decode logic to lib.rs
