+++ crates/challenger/src/challenger_processor.rs
@@ -70,29 +70,13 @@ where
 
     async fn handle_task_response(
         &mut self,
-        log: Log,
+        task_index: u32,
+        task_response: TaskResponse<TM::Output>,
+        task_response_metadata: TaskResponseMetadataSol,
+        non_signing_operator_pub_keys: Vec<G1Point>,
         is_response_correct: impl Fn(Task<TM::Input>, TaskResponse<TM::Output>) -> Result<bool, ChallengerError>
             + Send,
     ) -> Result<(), ChallengerError> {
-        let data = log.inner.data.data.0.clone();
-
-        // Decode a tuple of the form: (TaskResponse<TM::Output>, TaskResponseMetadata)
-        let ((task_index, response), task_response_metadata) =
-            <(
-                (
-                    <u32 as SolValue>::SolType,
-                    <TM::Output as SolValue>::SolType,
-                ),
-                <TaskResponseMetadataSol as SolValue>::SolType,
-            )>::abi_decode_params(&data, false)?;
-
-        let task_response = TaskResponse::<TM::Output> {
-            task_index,
-            response: response.into(),
-        };
-
-        let non_signing_operator_pub_keys = self.get_non_signing_operator_pub_keys(log).await?;
-
         if let Some(task) = self
             .tasks
             .get(&task_index)
@@ -135,51 +119,5 @@ where
             task_manager,
             tasks: HashMap::new(),
         }
-    }
-
-    pub async fn get_non_signing_operator_pub_keys(
-        &self,
-        log: Log,
-    ) -> Result<Vec<G1Point>, ChallengerError> {
-        let tx_hash = log
-            .transaction_hash
-            .ok_or(ChallengerError::TransactionHashNotFound)?;
-        let provider = get_provider(&self.rpc_url);
-
-        // TODO: Review this, rust-analyzer is not able to infer the type of the transaction
-        let tx: alloy::rpc::types::Transaction =
-            provider
-                .get_transaction_by_hash(tx_hash)
-                .await?
-                .ok_or(ChallengerError::TransactionNotFound(tx_hash.to_string()))?;
-
-        // The first 4 bytes are the selector, so we skip them
-        let calldata = tx
-            .inner
-            .input()
-            .get(4..)
-            .ok_or(ChallengerError::InvalidCalldata)?;
-
-        // Decode tuple of the form: Task<TM::Input>, TaskResponse<TM::Output>, NonSignerStakesAndSignature)
-        let (_, _, non_signer_stakes_and_signature) =
-            <(
-                (
-                    <TM::Input as SolValue>::SolType,
-                    <u32 as SolValue>::SolType,
-                    <Bytes as SolValue>::SolType,
-                    <u32 as SolValue>::SolType,
-                ),
-                (
-                    <u32 as SolValue>::SolType,
-                    <TM::Output as SolValue>::SolType,
-                ),
-                <NonSignerStakesAndSignature as SolValue>::SolType,
-            )>::abi_decode_params(calldata, false)?;
-
-        Ok(non_signer_stakes_and_signature
-            .nonSignerPubkeys
-            .into_iter()
-            .map(|pk| G1Point { X: pk.X, Y: pk.Y })
-            .collect())
     }
 }
