+++ crates/challenger/src/lib.rs
@@ -1,5 +1,6 @@
 #![allow(missing_docs)]
 use alloy::{
+    consensus::Transaction,
     dyn_abi::SolType,
     primitives::Bytes,
     providers::Provider,
@@ -7,8 +8,12 @@ use alloy::{
     sol_types::{SolEvent, SolValue},
 };
 use challenger::ChallengerTaskProcessor;
-use eigen_common::get_ws_provider;
+use challenger_processor::TaskResponseMetadataSol;
+use eigen_common::{get_provider, get_ws_provider};
 use eigen_task_processor::{task::Task, task_response::TaskResponse};
+use eigen_utils::slashing::middleware::iblssignaturechecker::{
+    IBLSSignatureCheckerTypes::NonSignerStakesAndSignature, BN254::G1Point,
+};
 use error::ChallengerError;
 use futures_util::StreamExt;
 use tracing::info;
@@ -21,6 +26,8 @@ pub mod task_manager;
 /// Main Challenger struct
 #[derive(Debug)]
 pub struct Challenger<TP: ChallengerTaskProcessor> {
+    /// The rpc url
+    rpc_url: String,
     /// The websocket url
     ws_url: String,
     /// The task processor
@@ -30,19 +37,22 @@ pub struct Challenger<TP: ChallengerTaskProcessor> {
 impl<TP: ChallengerTaskProcessor> Challenger<TP>
 where
     TP::Input: From<<<TP::Input as SolValue>::SolType as SolType>::RustType>,
+    TP::Output: From<<<TP::Output as SolValue>::SolType as SolType>::RustType>,
 {
     /// Create a new challenger
     ///
     /// # Arguments
     ///
+    /// * `rpc_url` - The rpc url
     /// * `ws_url` - The websocket url
     /// * `task_processor` - The task processor
     ///
     /// # Returns
     ///
     /// * `Self` - The challenger
-    pub fn new(ws_url: String, task_processor: TP) -> Self {
+    pub fn new(rpc_url: String, ws_url: String, task_processor: TP) -> Self {
         Self {
+            rpc_url,
             ws_url,
             task_processor,
         }
@@ -90,8 +100,16 @@ where
                     self.task_processor.handle_task_creation(task_index, task).await?;
                 },
                 Some(log) = responded_stream.next() => {
+                    let (task_index, task_response, task_response_metadata, non_signing_operator_pub_keys) =
+                        self.decode_task_response_event(log).await?;
                     self.task_processor
-                        .handle_task_response(log, &is_response_correct)
+                        .handle_task_response(
+                            task_index,
+                            task_response,
+                            task_response_metadata,
+                            non_signing_operator_pub_keys,
+                            &is_response_correct,
+                        )
                         .await?;
                 },
                 else => {
@@ -150,5 +168,90 @@ where
         };
 
         Ok((task_index, task))
+    }
+
+    async fn decode_task_response_event(
+        &self,
+        log: Log,
+    ) -> Result<
+        (
+            u32,
+            TaskResponse<TP::Output>,
+            TaskResponseMetadataSol,
+            Vec<G1Point>,
+        ),
+        ChallengerError,
+    > {
+        let data = log.inner.data.data.0.clone();
+
+        // Decode a tuple of the form: (TaskResponse<TM::Output>, TaskResponseMetadata)
+        let ((task_index, response), task_response_metadata) =
+            <(
+                (
+                    <u32 as SolValue>::SolType,
+                    <TP::Output as SolValue>::SolType,
+                ),
+                <TaskResponseMetadataSol as SolValue>::SolType,
+            )>::abi_decode_params(&data, false)?;
+
+        let task_response = TaskResponse::<TP::Output> {
+            task_index,
+            response: response.into(),
+        };
+
+        let non_signing_operator_pub_keys = self.get_non_signing_operator_pub_keys(log).await?;
+
+        Ok((
+            task_index,
+            task_response,
+            task_response_metadata,
+            non_signing_operator_pub_keys,
+        ))
+    }
+
+    pub async fn get_non_signing_operator_pub_keys(
+        &self,
+        log: Log,
+    ) -> Result<Vec<G1Point>, ChallengerError> {
+        let tx_hash = log
+            .transaction_hash
+            .ok_or(ChallengerError::TransactionHashNotFound)?;
+        let provider = get_provider(&self.rpc_url);
+
+        // TODO: Review this, rust-analyzer is not able to infer the type of the transaction
+        let tx: alloy::rpc::types::Transaction =
+            provider
+                .get_transaction_by_hash(tx_hash)
+                .await?
+                .ok_or(ChallengerError::TransactionNotFound(tx_hash.to_string()))?;
+
+        // The first 4 bytes are the selector, so we skip them
+        let calldata = tx
+            .inner
+            .input()
+            .get(4..)
+            .ok_or(ChallengerError::InvalidCalldata)?;
+
+        // Decode tuple of the form: Task<TM::Input>, TaskResponse<TM::Output>, NonSignerStakesAndSignature)
+        let (_, _, non_signer_stakes_and_signature) =
+            <(
+                (
+                    <TP::Input as SolValue>::SolType,
+                    <u32 as SolValue>::SolType,
+                    <Bytes as SolValue>::SolType,
+                    <u32 as SolValue>::SolType,
+                ),
+                (
+                    <u32 as SolValue>::SolType,
+                    <TP::Output as SolValue>::SolType,
+                ),
+                <NonSignerStakesAndSignature as SolValue>::SolType,
+            )>::abi_decode_params(calldata, false)?;
+
+        Ok(non_signer_stakes_and_signature
+            .nonSignerPubkeys
+            .into_iter()
+            .map(|pk| G1Point { X: pk.X, Y: pk.Y })
+            .collect())
     }
 }
